Split up tests into packages
