+++ subprojects/logging/src/integTest/groovy/org/gradle/internal/logging/console/taskgrouping/ConsoleBuildSrcGroupedTaskFunctionalTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.logging.console
+package org.gradle.internal.logging.console.taskgrouping
 
 class ConsoleBuildSrcGroupedTaskFunctionalTest extends AbstractConsoleGroupedTaskFunctionalTest {
