+++ subprojects/logging/src/integTest/groovy/org/gradle/internal/logging/console/taskgrouping/ConsoleGradleBuildGroupedTaskFunctionalTest.groovy
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.gradle.internal.logging.console
+package org.gradle.internal.logging.console.taskgrouping
 
 import spock.lang.FailsWith
